# ficha8/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Ficha 8 API",
        "description": "Ficha 8 API Information",
        "contact": {
            "name": "TPSI-DWB"
        },
        "servers": [
            "http://localhost:3000"
        ]
    },
    "paths": {
        "/person":{
            "post":{
                "tags":[
                    "person"
                ],
                "summary":"Create person",
                "description":"This can only be done by the logged in user",
                "operationId":"createPerson",
                "produces":[
                    "application/xml",
                    "application/json"
                ],
                "parameters":[
                    {
                        "in":"body",
                        "name":"body",
                        "description":"Create person object",
                        "required":true,
                        "schema":{
                            "$ref":"#/definitions/Person"
                        }
                    }
                ],
                "responses":{
                    "default":{
                        "description":"Successful operation"
                    }
                }
            }
        },
        "/person/{id}":{
            "get":{
                "tags":[
                    "person"
                ],
                "summary":"Get person by id",
                "description":"Get person by id",
                "operationId":"getPersonById",
                "produces":[
                    "application/xml",
                    "application/json"
                ],
                "parameters":[
                    {
                        "in":"path",
                        "name":"id",
                        "description":"The id that needs to be fetched. Use id 1 for testing",
                        "required":true,
                        "type":"string",
                        "schema":{
                            "$ref":"#/definitions/Person"
                        }
                    }
                ],
                "responses":{
                    "200":{
                        "description":"Successful operation",
                        "schema":{
                            "$ref":"#/definitions/Person"
                        }
                    },
                    "400":{
                        "description":"Invalid id supplied"
                    },
                    "404":{
                        "description":"Id not found"
                    }
                }
            }
        }
    },
    "definitions": {
        "Person": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "x-primary-key": true
                },
                "firstname": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "profession": {
                    "type": "string"
                },
                "age": {
                    "type": "integer",
                    "format": "int64"
                }
            }
        }
    }
}